+++ CHANGELOG.md
@@ -1,7 +1,8 @@
-# isomorphic-lib-template Changelog
+# transaction-manager-service Changelog
 
 ## 1.0.0 - TBD
 
 ### Added
 
 - Initial commit.
+- First working MVP commit.